refactor: hooks called in parent + MenuManager
