+++ components/navbar/DesktopMenu.js
@@ -1,25 +1,7 @@
-import { Box, Button, Skeleton } from "@mui/material";
+import { Box, Button } from "@mui/material";
 import Link from "next/link";
-import useCurriculums from "../../hooks/use-curriculums";
 
-function DesktopMenu({ loggedIn }) {
-  const pages = useCurriculums();
-
-  if (pages === undefined) {
-    return (
-      <Skeleton
-        variant="rounded"
-        width={10}
-        height={30}
-        sx={{
-          flexGrow: 1,
-          display: { xs: "none", md: "flex" },
-          mr: 75,
-        }}
-      />
-    );
-  }
-
+function DesktopMenu({ loggedIn, pages }) {
   return (
     <Box
       sx={{
