+++ components/navbar/Navbar.js
@@ -1,12 +1,13 @@
 import { AppBar, Toolbar, Container } from "@mui/material";
 import AppTitleAndLogo from "./AppTitleAndLogo";
 import UserBubble from "./UserBubble";
-import DesktopMenu from "./DesktopMenu";
-import MobileMenu from "./MobileMenu";
 import useCurrentUser from "../../modules/authentication/hooks/use-current-user";
+import useCurriculums from "../../hooks/use-curriculums";
+import MenuManager from "./MenuManager";
 
 function Navbar() {
   const currentUser = useCurrentUser();
+  const pages = useCurriculums();
   return (
     <AppBar position="static">
       <Container maxWidth="xl">
@@ -16,13 +17,21 @@ function Navbar() {
             variant="h6"
             flexGrow={0}
           />
-          <MobileMenu loggedIn={!!currentUser} />
+          <MenuManager
+            version="mobile"
+            loggedIn={!!currentUser}
+            pages={pages}
+          />
           <AppTitleAndLogo
             display={{ xs: "flex", md: "none" }}
             variant="h5"
             flexGrow={1}
           />
-          <DesktopMenu loggedIn={!!currentUser} />
+          <MenuManager
+            version="desktop"
+            loggedIn={!!currentUser}
+            pages={pages}
+          />
           <UserBubble />
         </Toolbar>
       </Container>
